// src/app/category/[slug]/components/interfaces.ts
export interface BannerImage {
    id: string;
    image?: { full: string };
    url?: string;
    link?: string;
    order: number;
}

export interface CategoryBannerProps {
    category: CategoryData | null;
    bannerData?: { images: BannerImage[] };
}